Minimally-Viable Minimal Viable Prototype successfully implemented
The current version of the code base successfully implements a
prototype version which is functional, as far as the actual
generation of basic problems is concerned.

The IntegrationProblem class was slapped together in a very
haphazard way, and proper sub-classing and API design will be
necessary to get to a happy place from here.